remote_agent/requirements.txt
google-adk
google-adk[eval]
google-adk[A2A]
pydantic
yfinance
yahoo-finance-server
